chore: disable CS3021 warning for the antlr generated code
Summary:

In the generated code from the antlr code has this warning in it. Rather than
try and fix it, this disables the messages.

Because this is generated code making any changes to it becomes a maintenance
issue.

Test Plan:

ATM there is no test plan. Soon we will add the warn as error flag to the
compiler that will ensure we are not emitting any warnings.