fix: ensure the falsy condition in the if dose not get evaluated
Summary:

We we are doing if else statements we are evaluate both truthy and falsy
condition. This obviously not what we want.

Now we are making sure we only evaluate the list in the condition and the
correct side of the if else. This is the last bug that needed fixing before we
can add the fibonacci test.

We now support recursive functions :)

Test Plan:

Unit tests