fix: passing a list into a function as a variable
Summary:

When you pass a list in to a function as a variable is is not getting
evaluated. For example in the below code:

```lisp
(defun plus-one (a) (+ 1 a))
(plus-one (- 1 2))
```

The parameter to `plus-one` is being pass in as a list not the number `2`. This
is then causing an error in the `+` function where the variable a is a list and
not a number.

Now we are evaluating the list when we set the variable `a`, this is giving you
the expected behaviour of `2`.

Test Plan:

Unit tests.