feat: improve error messages
Summary:

Now when you get an error you get a location and annotated code of the error.
Previously when you got an error, you had a message but that was it. This made
debugging quite hard even for the small programs in the test suite.

This is an example of the error you will get now.

```
./TinyLisp.Integration/basic.lsp(4,9): Symbol `==` is not defined in the current scope

   2 |
   3 | ;; Test we can do some maths
   4 | (assert (== (+ 1 1) 2) "One plus one must be 2")
                ^^ Symbol `==` is not defined in the current scope
   5 | (assert (= (- 1 1) 0) "One minus one must be 0")
   6 | (assert (= (/ 10 2) 5) "This must be 5")
```

Test Plan:

N / A. This has been tested manually